+++ vistle/core/scalar.h
@@ -3,7 +3,7 @@
 
 namespace vistle {
 
-typedef double Scalar;
+typedef float Scalar;
 
 } // namespace vistle
